requirements.txt
python-dotenv
pandas
plotly
requests
playwright
